Fixed several warnings; Implemented deletion of annations;
git-svn-id: https://svn.code.sf.net/p/podofo/code/podofo/trunk@304 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5

+++ src/PdfDocument.cpp
@@ -658,7 +658,7 @@ bool PdfDocument::AddNamedDestination( PdfDestination& inDest, const std::string
                 }
             } else {	// are replacing
                 size_t idx = destArray.GetStringIndex( inName );
-                if ( idx == -1 ) {	// not found, so put it at the end
+                if ( static_cast<int>(idx) == -1 ) {	// not found, so put it at the end
                     destArray.push_back( PdfString( inName ) );
                     destArray.push_back( *inDest.GetObject() );
                 } else {
